Move GridViewEditAccessibleObject to own file
